+++ cypress/pageObjects/IconLegendsPage.js
@@ -5,7 +5,6 @@ class IconLegends {
     getProjectHealthGroupHeader = () => cy.get('#main-panel>h2:nth-child(5)');
     getProjectHealthIconsGroup = () => cy.get('#main-panel > dl:nth-child(6) dt');
 
-    
 }
 
 export default IconLegends;
